Move swipe down file to components folder
